+++ h2h/user/src/index.html
@@ -25,6 +25,7 @@
 					<li>
 						<a href="login.html">登录</a>
 					</li>
+					<span></span>
 					<li>
 						<a href="register.html">注册</a>
 					</li>
@@ -44,7 +45,7 @@
 								<!--库存-->
 								<span class="name">剩余<span id = "stock1"></span>本</span>
 								<!--租金-->
-								<div class="row">租金： 元 <span id = "price1"></span> /天</div>					
+								<div class="row">租金：  <span id = "price1"></span> 元/天</div>					
 							</div>	
 						</div>
 						<div class="box-overlay magictime inner" data-hover="rotateLeft" data-return="">
@@ -71,7 +72,7 @@
 								<!--库存-->
 								<span class="name">剩余<span id = "stock2"></span>本</span>
 								<!--租金-->
-								<div class="row">租金： 元 <span id = "price2"></span> /天</div>						
+								<div class="row">租金：  <span id = "price2"></span> 元/天</div>						
 							</div>	
 						</div>
 						<div class="box-overlay magictime inner" data-hover="rotateLeft" data-return="">
@@ -97,7 +98,7 @@
 								<!--库存-->
 								<span class="name">剩余：<span id = "stock3"></span>本</span>
 								<!--租金-->
-								<div class="row">租金： 元 <span id = "price3"></span> /天</div>						
+								<div class="row">租金：  <span id = "price3"></span>元 /天</div>						
 							</div>	
 						</div>
 						<div class="box-overlay magictime inner" data-hover="rotateLeft" data-return="">
@@ -125,7 +126,7 @@
 								<!--库存-->
 								<span class="name">剩余<span id = "stock4"></span>本</span>
 								<!--租金-->
-								<div class="row">租金： 元 <span id = "price4"></span> /天</div>						
+								<div class="row">租金： <span id = "price4"></span> 元/天</div>						
 							</div>	
 						</div>
 						<div class="box-overlay magictime inner" data-hover="rotateLeft" data-return="">
@@ -152,7 +153,7 @@
 								<!--库存-->
 								<span class="name">剩余<span id = "stock5"></span>本</span>
 								<!--租金-->
-								<div class="row">租金： 元 <span id = "price5"></span> /天</div>						
+								<div class="row">租金： <span id = "price5"></span>元 /天</div>						
 							</div>	
 						</div>
 						<div class="box-overlay magictime inner" data-hover="rotateLeft" data-return="">
@@ -179,7 +180,7 @@
 								<!--库存-->
 								<span class="name">剩余<span id = "stock6"></span>本</span>
 								<!--租金-->
-								<div class="row">租金： 元 <span id = "price6"></span> /天</div>						
+								<div class="row">租金： <span id = "price6"></span>元 /天</div>						
 							</div>	
 						</div>
 						<div class="box-overlay magictime inner" data-hover="rotateLeft" data-return="">
@@ -206,7 +207,7 @@
 								<!--库存-->
 								<span class="name">剩余<span id = "stock7"></span>本</span>
 								<!--租金-->
-								<div class="row">租金： 元 <span id = "price7"></span> /天</div>							
+								<div class="row">租金： <span id = "price7"></span>元 /天</div>							
 							</div>	
 						</div>
 						<div class="box-overlay magictime inner" data-hover="rotateLeft" data-return="">
@@ -234,7 +235,7 @@
 								<!--库存-->
 								<span class="name">剩余<span id = "stock8"></span>本</span>
 								<!--租金-->
-								<div class="row">租金：元<span id = "price8"></span>/天</div>						
+								<div class="row">租金：<span id = "price8"></span>元/天</div>						
 							</div>	
 						</div>
 						<div class="box-overlay magictime inner" data-hover="rotateLeft" data-return="">
@@ -284,20 +285,21 @@
 	<script src="js/vender/jquery.1.10.2.min.js"></script>
 	<script src="js/vender/jquery.cookie.js"></script>
     <script src="js/index/index.min.js"></script>
+    <script src="js/vender/cookie.js"></script>
 </body>
 <script type="text/javascript">
 	//var bookIds = new Array();
 	window.onload = function() {
 		
 		//获取登录cookie
-		if($.cookie("login") == null){
+		if(getCookie("login") == null){
 			
 		}
-		if($.cookie("login") == "true"){
-			$('#title ul li:nth-child(1) a').text($.cookie("email"));
+		if(getCookie("login") == "true"){
+			$('#title ul li:nth-child(1) a').text(getCookie("email"));
 			$('#title ul li:nth-child(1) a').attr("href","#");
-			$('#title ul li:nth-child(2) a').text("退出");
-			$('#title ul li:nth-child(2) a').attr("href","#");
+			$('#title ul li:nth-child(3) a').text("退出");
+			$('#title ul li:nth-child(3) a').attr("href","#");
 			
 		}
