+++ h2h/user/src/logic/addToCart.php
@@ -1,9 +1,9 @@
 <?php
 require_once dirname(__FILE__)."/../../../db/DBCart.php";
-$email = $_POST['useremail'];
+$email = $_POST['email'];
 $bookid = $_POST['bookid'];
-//$email = "13301054@bjtu.edu.cn";
-//$bookid = 00000004;
+// $email = "13301054@bjtu.edu.cn";
+// $bookid = "00000001";
 $dbcart = new DBCart();
 
 $add = $dbcart->addToCart($email,$bookid);
